+++ gazebo/physics/Collision.cc
@@ -201,7 +201,7 @@ ShapePtr Collision::GetShape() const
 //////////////////////////////////////////////////
 void Collision::SetScale(const math::Vector3 &_scale)
 {
-  this->shape->SetScale(_scale);
+  this->shape->SetScale(_scale.Ign());
 }
 
 //////////////////////////////////////////////////